requirements.txt
anyio==4.4.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-telegram-bot==21.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
urllib3==2.2.2
